+++ packages/camera/camera/android/src/main/java/io/flutter/plugins/camera/PictureCaptureRequest.java
@@ -7,7 +7,9 @@ class PictureCaptureRequest {
 
   enum State {
     idle,
-    awaitingPreCapture,
+    focusing,
+    preCapture,
+    waitingPreCaptureReady,
     capturing,
     finished,
     error,