+++ packages/camera/camera/lib/camera.dart
@@ -12,6 +12,7 @@ export 'package:camera_platform_interface/camera_platform_interface.dart'
         CameraException,
         CameraLensDirection,
         FlashMode,
+        ExposureMode,
         ResolutionPreset,
         XFile,
         ImageFormatGroup;